Clean up example project
* Moved all code to lib
* Made named area Piet painting the main example
* Hoisted explicitly placed Piet painting directly into test
* Fixed up all links in README
