--- example/flutter_layout_grid.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-import './piet_painting.dart';
-
-void main() {
-  // Checkout out piet_painting.dart
-  runApp(PietApp());
-}
